+++ src/app/civil-servant/components/paystub-links/index.ts
@@ -1,2 +1,0 @@
-export * from './paystub-links.component.module'
-export * from './paystub-links.component'