+++ src/app/civil-servant/components/report-yields-download-form/report-yields-download-form.component.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { IonicModule } from 'ionic-angular';
+import { ValidationMessageModule } from '@espm/shared';
+import { ReportYieldsDownloadFormComponent } from './report-yields-download-form.component'
+
+@NgModule({
+  declarations: [ReportYieldsDownloadFormComponent],
+  imports: [
+    CommonModule,
+    IonicModule,
+    ValidationMessageModule
+  ],
+  exports: [ReportYieldsDownloadFormComponent],
+  providers: [],
+})
+export class ReportYieldsDownloadFormModule { }