+++ src/app/civil-servant/components/report-yields-download/report-yields-download.component.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { IonicPageModule } from 'ionic-angular';
+import { ReportYieldsDownloadComponent } from './report-yields-download.component';
+import { ModulePageComponentModule } from '@espm/shared';
+import { ReportYieldsDownloadFormModule } from '../report-yields-download-form';
+
+@NgModule({
+  declarations: [
+    ReportYieldsDownloadComponent,
+  ],
+  imports: [
+    ModulePageComponentModule,
+    ReportYieldsDownloadFormModule,
+    IonicPageModule,
+  ],
+  exports: [
+    ReportYieldsDownloadComponent,
+  ]
+})
+export class ReportYieldsDownloadComponentModule { }