+++ src/app/civil-servant/components/siarhes-profiles/siarhes-profiles.component.ts
@@ -1,15 +1,15 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { NavController, NavParams, /*AlertController*/ } from 'ionic-angular';
-import { IPaystubProfile } from '../../interfaces/profile';
+import { ISiarhesProfile } from '../../interfaces/profile';
 
 @Component({
-  selector: 'paystub-profiles',
-  templateUrl: 'paystub-profiles.component.html'
+  selector: 'siarhes-profiles',
+  templateUrl: 'siarhes-profiles.component.html'
 })
-export class PaystubProfilesComponent {
+export class SiarhesProfilesComponent {
 
-  @Input() profiles: IPaystubProfile[] = [];
-  @Output() onSelectProfile: EventEmitter<IPaystubProfile> = new EventEmitter();
+  @Input() profiles: ISiarhesProfile[] = [];
+  @Output() onSelectProfile: EventEmitter<ISiarhesProfile> = new EventEmitter();
 
   constructor(public navCtrl: NavController, public navParams: NavParams, /*private alert: AlertController*/) { }
 
@@ -24,7 +24,7 @@ export class PaystubProfilesComponent {
   //   alert.present();
   // }
 
-  selectProfile(profile: IPaystubProfile) {
+  selectProfile(profile: ISiarhesProfile) {
     this.onSelectProfile.emit(profile);
   }
 }